refactor: organize by folder for each registry file
