+++ src/aibtc-dao/extensions/core-proposals/deploy-core-proposal.ts
@@ -1,7 +1,7 @@
 import * as fs from "fs";
 import * as path from "path";
-import { getContractName } from "../../registry/dao-contract-registry";
-import { DeployedCoreProposalRegistryEntry } from "../../registry/dao-core-proposal-registry";
+import { getContractName } from "../../registries/dao-contract-registry";
+import { DeployedCoreProposalRegistryEntry } from "../../registries/dao-core-proposal-registry";
 import { DaoCoreProposalGenerator } from "../../services/dao-core-proposal-generator";
 import { DaoCoreProposalDeployer } from "../../services/dao-core-proposal-deployer";
 import {
