+++ src/aibtc-dao/services/dao-core-proposal-generator.ts
@@ -5,11 +5,11 @@ import { StacksNetworkName } from "@stacks/network";
 import {
   CORE_PROPOSAL_REGISTRY,
   GeneratedCoreProposalRegistryEntry,
-} from "../registry/dao-core-proposal-registry";
+} from "../registries/dao-core-proposal-registry";
 import {
   getContractName,
   getContractsBySubcategory,
-} from "../registry/dao-contract-registry";
+} from "../registries/dao-contract-registry";
 import { getKnownAddresses, getKnownTraits } from "../types/dao-types";
 
 export class DaoCoreProposalGenerator {
